Move IGNORED_ERRORS to its own file
We're going to need to import this in multiple places.
